jpql

+++ src/main/java/com/example/ecom_proj/controller/ProductController.java
@@ -19,6 +19,9 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.RequestPart;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.http.MediaType;
+import org.springframework.web.bind.annotation.RequestParam;
+
 
 
 @RestController
@@ -55,9 +58,41 @@ public class ProductController {
         }
     }
 
+    @GetMapping("/products/{id}/image")
+    public ResponseEntity<byte[]> getImageByProductId(@PathVariable int id) {
+        Product product = service.getProductById(id);
+        byte[] image = product.getImageData();
+
+        return ResponseEntity.ok().contentType(MediaType.valueOf(product.getImageType()))
+                .body(image);
+        
+    }
+
+    @PutMapping("/products/{id}")
+    public ResponseEntity<?> updateProduct(@PathVariable int id, @RequestPart Product product,@RequestPart MultipartFile image) {
+        try{
+        Product updatedProduct = service.updateProduct(id, product,image);
+        return new ResponseEntity<>(updatedProduct, HttpStatus.OK);
+        }catch (Exception e){
+            return new ResponseEntity<>(e.getMessage() ,HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+    }
+
     @DeleteMapping("/products/{id}")
-    public void deleteProductById(@PathVariable int id) {
+    public ResponseEntity<?> deleteProductById(@PathVariable int id) {
+        Product product = service.getProductById(id);
+        if(product == null){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         service.deleteProductById(id);
+        return new ResponseEntity<>(HttpStatus.OK);
+    }
+
+    @GetMapping("/products/search")
+    public ResponseEntity<List<Product>> searchProducts(@RequestParam String keyword) {
+        System.out.println("Keyword: " + keyword);
+        List<Product> products = service.searchProducts(keyword);
+        return new ResponseEntity<>(products, HttpStatus.OK);
     }
 
 }
